Ação de salvar e alterar produto

+++ app/app/src/main/java/com/alura/sida/utils/NumberUtils.java
@@ -3,7 +3,7 @@ package com.alura.sida.utils;
 import java.text.DecimalFormat;
 
 public class NumberUtils {
-    public static String  formatarDecimal(Double valor)
+    public static String  formatarDecimal(float valor)
     {
         String resultado;
         DecimalFormat formato = new DecimalFormat("#.00");
